(mode_t) [_WIN32]: Switch to unsigned short to match the declaration in the MinGW for mode_t.

+++ src/lib/w32/w32misc.h
@@ -52,7 +52,7 @@
 #define srandom srand
 
 /* sys/types.h */
-typedef int mode_t;
+typedef unsigned short mode_t;
 typedef long off_t;
 typedef int pid_t;
 typedef long ssize_t;
